2013-03-23: Merged fixes (explicit type casts) from 'memisc-0.96' into 'codebooks'
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/memisc@34 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/codebooks/src/readfixed.c
@@ -11,15 +11,15 @@
 
 int _R_atoi(char *text){
   char *end_ptr;
-  if(!strlen(text)) return NA_INTEGER;
-  int result = strtol(text,&end_ptr,10);
+  if(!strlen(text)) return NA_INTEGER; 
+  int result = (int)strtol(text,&end_ptr,10);
   if(!isdigit(end_ptr[-1])) return NA_INTEGER;
   else return result;
 }
 
 double _R_atof(char *text){
   char *end_ptr;
-  if(!strlen(text)) return NA_REAL;
+  if(!strlen(text)) return NA_REAL; 
   double result = strtod(text,&end_ptr);
   if(!isdigit(end_ptr[-1]) && isdigit(end_ptr[-1]) != '.') return NA_REAL;
   else return result;
